cancellazione??????

+++ src/ats/controllo/servlet/DeleteServlet.java
@@ -25,6 +25,7 @@ public class DeleteServlet extends HttpServlet {
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)	throws ServletException, IOException {
 		Long id = Long.parseLong(request.getParameter("id"));
+
 		IDAOUtente daoUtente = new DAOUtente();
 		try {
 			Utente u = daoUtente.findById(id);
@@ -39,7 +40,7 @@ public class DeleteServlet extends HttpServlet {
 			System.out.println(e.getMessage());
 		}
 		
-		if (id == (Long) request.getSession().getAttribute("id")) {
+		if (id.longValue() == ((Long) request.getSession().getAttribute("id")).longValue()) {
 			request.getSession().invalidate();
 			response.sendRedirect("/GitDAOTaxi/index.jsp");
 		} else {
